# openapi-schema.json
{"openapi":"3.0.0","paths":{"/api/v1/users":{"get":{"operationId":"UserController_getUsers","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetUserDto"}}}}}},"tags":["USER"]}},"/api/v1/users/{id}":{"get":{"operationId":"UserController_getUserById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserDto"}}}}},"tags":["USER"]}},"/api/v1/users/auth/signup":{"post":{"operationId":"UserController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserDto"}}}}},"tags":["USER"]}},"/api/v1/users/auth/signin":{"post":{"operationId":"UserController_signin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninResponseDto"}}}}},"tags":["USER"]}},"/api/v1/users/update/{id}":{"put":{"operationId":"UserController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserDto"}}}}},"tags":["USER"]}},"/api/v1/location/departments/insert-data":{"get":{"operationId":"DepartmentController_insertData","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["LOCATIONS / DEPARTMENTS"]}},"/api/v1/location/departments/all":{"get":{"operationId":"DepartmentController_getDepartments","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}},"tags":["LOCATIONS / DEPARTMENTS"]}},"/api/v1/location/departments/{depName}":{"get":{"operationId":"DepartmentController_getDepartmentByName","parameters":[{"name":"depName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetDepartmentByNameDto"}}}}},"tags":["LOCATIONS / DEPARTMENTS"]}},"/api/v1/location/departments/{depName}/towns":{"get":{"operationId":"DepartmentController_getTowns","parameters":[{"name":"depName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetDepartmentTownsDto"}}}}},"tags":["LOCATIONS / DEPARTMENTS"]}},"/api/v1/location/departments/{depName}/districts":{"get":{"operationId":"DepartmentController_getDistricts","parameters":[{"name":"depName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetDepartmentDistrictsDto"}}}}},"tags":["LOCATIONS / DEPARTMENTS"]}},"/api/v1/location/departments/{depName}/neighborhoods":{"get":{"operationId":"DepartmentController_getNeighborhoods","parameters":[{"name":"depName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetDepartmentNeighborhoodsDto"}}}}},"tags":["LOCATIONS / DEPARTMENTS"]}},"/api/v1/location/districts/all":{"get":{"operationId":"DistrictController_getDistricts","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}},"tags":["LOCATION / DISTRICTS"]}},"/api/v1/location/districts/{districtName}/neighborhoods":{"get":{"operationId":"DistrictController_getNeighborhoods","parameters":[{"name":"districtName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetDistrictNeighborhoodsDto"}}}}},"tags":["LOCATION / DISTRICTS"]}},"/api/v1/location/towns/all":{"get":{"operationId":"TownController_getTowns","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetTownDto"}}}}}},"tags":["LOCATION / TOWNS"]}},"/api/v1/location/towns/{townName}/districts":{"get":{"operationId":"TownController_getDistricts","parameters":[{"name":"townName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetTownDistrictsDto"}}}}},"tags":["LOCATION / TOWNS"]}},"/api/v1/location/towns/{townName}/neighborhoods":{"get":{"operationId":"TownController_getNeighborhoods","parameters":[{"name":"townName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetTownNeighborhoodsDto"}}}}},"tags":["LOCATION / TOWNS"]}},"/api/v1/location/neighborhoods":{"post":{"operationId":"NeighborhoodController_getNeighborhoods","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginateDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetNeighborhoods"}}}}},"tags":["LOCATION / NEIGHBORHOODS"]}}},"info":{"title":"BJ LOCATIONS REST API","description":"This is the documentation for the BJ LOCATIONS REST API","version":"1.0","contact":{"name":"THE NERONE TEAM","url":"https://www.dev.metatronicx.com","email":"noreply@metatronicx.com"}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"GetUserDto":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"createdAt":{"type":"string"}},"required":["id","email","createdAt"]},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email"]},"AuthenticateUserDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"SigninResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"email":{"type":"string"}},"required":["accessToken","email"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email"]},"GetDepartmentByNameDto":{"type":"object","properties":{"name":{"type":"string"},"createdAt":{"type":"string"}},"required":["name","createdAt"]},"GetDepartmentTownsDto":{"type":"object","properties":{"department":{"type":"string"},"towns":{"type":"array","items":{"type":"string"}}},"required":["department","towns"]},"GetDepartmentDistrictsDto":{"type":"object","properties":{"department":{"type":"string"},"districts":{"type":"array","items":{"type":"string"}}},"required":["department","districts"]},"GetDepartmentNeighborhoodsDto":{"type":"object","properties":{"department":{"type":"string"},"neighborhoods":{"type":"array","items":{"type":"string"}}},"required":["department","neighborhoods"]},"GetDistrictNeighborhoodsDto":{"type":"object","properties":{"district":{"type":"string"},"neighborhoods":{"type":"array","items":{"type":"string"}}},"required":["district","neighborhoods"]},"GetTownDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"}},"required":["id","name"]},"GetTownDistrictsDto":{"type":"object","properties":{"town":{"type":"string"},"districts":{"type":"array","items":{"type":"string"}}},"required":["town","districts"]},"GetTownNeighborhoodsDto":{"type":"object","properties":{"town":{"type":"string"},"neighborhoods":{"type":"array","items":{"type":"string"}}},"required":["town","neighborhoods"]},"PaginateDto":{"type":"object","properties":{"pageNumber":{"type":"number"},"dataLength":{"type":"number"}},"required":["pageNumber"]},"GetNeighborhoods":{"type":"object","properties":{"data":{"type":"array","items":{"type":"string"}},"dataCount":{"type":"number"},"totalRecordsCount":{"type":"number"},"pageNumbers":{"type":"array","items":{"type":"string"}}},"required":["data","dataCount","totalRecordsCount","pageNumbers"]}}}}